Globalized Sakura Kensha Functions
