+++ .gitignore
@@ -1,5 +1,3 @@
 env
 __pycache__
 data/*
-
-!data/.gitkeep
