+++ parsers/sakura_kensha_parser.py
@@ -1,9 +1,7 @@
 from bs4 import BeautifulSoup
 import re
 import dateparser
-import json
 from pathlib import Path
-
 
 from .base_parser import BaseParser
 from utils.fetch import fetch_html
@@ -11,16 +9,21 @@ from models.sakura_kensha_model import LitterInfo
 
 
 class SakuraKenshaParser(BaseParser):
-    def __init__(self, url):
+    def __init__(self, url, storage_path="data/sakura_kensha.json"):
+        self.name = "Sakura Kensha"
+        self.storage_path = Path(storage_path)
         self.url = url
         self.html = fetch_html(self.url)
         self.litters = self.parse(self.html)
-        self.storage_path = Path("data/sakura_kensha.json")
 
-        self.previous_litters = self.load_previous_litters()
-        self.check_for_changes()
-        self.save_litters()
-        # self.display_litters()
+        new_litters = self.check_for_changes()
+
+        if new_litters:
+            self.send_notification(new_litters)
+        else:
+            print("Pas de changement.")
+
+        # self.display_litter()
 
     def parse(self, html):
         # Parse the html and return a list of elements
@@ -28,7 +31,7 @@ class SakuraKenshaParser(BaseParser):
         contents = soup.find_all(class_="elementor-widget-text-editor")
         elements = [content.get_text(strip=True) for content in contents]
 
-        # Parse the elements into a list of LitterInfo objects (the )
+        # Parse the elements into a list of LitterInfo objects
         litters = []
         # All the informations of a litter is in 3 consecutive elements
         for i in range(0, len(elements), 3):
@@ -69,25 +72,3 @@ class SakuraKenshaParser(BaseParser):
             except Exception as e:
                 print(f"Error parsing litter from Sakura Kensha: {e}")
         return litters
-
-    def display_litters(self):
-        for litter in self.litters:
-            print(litter)
-
-    def save_litters(self):
-        with open(self.storage_path, "w", encoding="utf-8") as f:
-            json.dump([litter.to_dict() for litter in self.litters],
-                      f, ensure_ascii=False, indent=4, default=str)
-
-    def load_previous_litters(self):
-        if self.storage_path.exists():
-            with open(self.storage_path, "r", encoding="utf-8") as f:
-                return json.load(f)
-        return []
-
-    def check_for_changes(self):
-        current_litters = [litter.to_dict() for litter in self.litters]
-        if current_litters != self.previous_litters:
-            print("Il y a eu un changement !")
-        else:
-            print("Pas de changement.")
